fix request receipt queries dfor substrate in hyperclient
