+++ modules/client/Cargo.toml
@@ -22,7 +22,6 @@ codec = { package = "parity-scale-codec", version = "3.1.3", default-features = 
 futures = "0.3.30"
 wasm-streams = "0.4.0"
 tokio = { version = "1.35.1", features = ["macros"]  }
-wasm-bindgen-test = "0.3.41"
 js-sys = "0.3.68"
 web-sys = "0.3.68"
 wasm-timer = { package = "fluvio-wasm-timer", version = "0.2.5" }
@@ -41,6 +40,7 @@ evm-common = { workspace = true }
 ismp-solidity-abi = { workspace = true }
 ethereum-trie = { workspace = true }
 subxt-utils = { workspace = true }
+hex = "0.4.3"
 
 [dependencies.reconnecting-jsonrpsee-ws-client]
 git = "https://github.com/niklasad1/reconnecting-jsonrpsee-ws-client"
@@ -51,8 +51,15 @@ default-features = false
 default = ["std"]
 wasm = ["subxt/web", "subxt/jsonrpsee", "reconnecting-jsonrpsee-ws-client/web", "subxt-utils/wasm"]
 std = ["subxt/native", "subxt/jsonrpsee", "reconnecting-jsonrpsee-ws-client/native", "subxt-utils/std"]
+testing = []
 
 [dev-dependencies]
 json = { package = "serde_json", version = "1.0.114", default-features = false }
 wasm-bindgen-test = "0.3.24"
 hex = { version = "0.4.3", default-features = false }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
+ethers = { workspace = true, features = ["ws"], default-features = true }
+pallet-ismp = { workspace = true, default-features = true }
+tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
+tokio = { version = "1.35.1", features = ["macros", "rt-multi-thread"]  }
